Fixed the tests
* Fixed the specs 2 issue where failures were silent
* Fixed ordering issues with Flash tests
* Moved some integration style tests into the integration project
* Fixed failing tests that weren't reporting before
