+++ framework/src/play-test/src/main/scala/play/api/test/Specs.scala
@@ -17,7 +17,7 @@ import org.specs2.execute.{ AsResult, Result }
 abstract class WithApplication(val app: FakeApplication = FakeApplication()) extends Around with Scope {
   implicit def implicitApp = app
   override def around[T: AsResult](t: => T): Result = {
-    Helpers.running(app)(AsResult(t))
+    Helpers.running(app)(AsResult.effectively(t))
   }
 }
 
@@ -54,7 +54,7 @@ abstract class WithBrowser[WEBDRIVER <: WebDriver](
 
   override def around[T: AsResult](t: => T): Result = {
     try {
-      Helpers.running(TestServer(port, app))(AsResult(t))
+      Helpers.running(TestServer(port, app))(AsResult.effectively(t))
     } finally {
       browser.quit()
     }
