+++ framework/src/play/src/test/scala/play/api/mvc/ResultsSpec.scala
@@ -1,8 +1,6 @@
 package play.api.mvc
 
 import org.specs2.mutable._
-import org.specs2.specification.Scope
-import org.specs2.execute.{Result => SpecsResult,AsResult}
 import play.api.libs.iteratee.{Iteratee, Enumerator}
 import scala.concurrent.Await
 import scala.concurrent.duration.Duration
@@ -77,105 +75,6 @@ object ResultsSpec extends Specification {
     "allow discarding multiple cookies by deprecated names method" in {
       val cookies = Cookies.decode(Ok.discardingCookies("foo", "bar").header.headers("Set-Cookie")).map(_.name)
       cookies must containTheSameElementsAs(Seq("foo", "bar"))
-    }
-
-    "support session helper" in new WithApplication {
-
-      Session.decode("  ").isEmpty must be_==(true)
-
-      val data = Map("user" -> "kiki", "bad:key" -> "yop", "langs" -> "fr:en:de")
-      val encodedSession = Session.encode(data)
-      val decodedSession = Session.decode(encodedSession)
-
-      decodedSession.size must be_==(2)
-      decodedSession must havePair("user" -> "kiki")
-      decodedSession must havePair("langs" -> "fr:en:de")
-      val SimpleResult(ResponseHeader(_, headers), _, _) =
-        Ok("hello").as("text/html")
-          .withSession("user" -> "kiki", "langs" -> "fr:en:de")
-          .withCookies(Cookie("session", "items"), Cookie("preferences", "blue"))
-          .discardingCookies(DiscardingCookie("logged"))
-          .withSession("user" -> "kiki", "langs" -> "fr:en:de")
-          .withCookies(Cookie("lang", "fr"), Cookie("session", "items2"))
-
-      val setCookies = Cookies.decode(headers("Set-Cookie")).map(c => c.name -> c).toMap
-      setCookies.size must be_==(5)
-      setCookies("session").value must be_==("items2")
-      setCookies("preferences").value must be_==("blue")
-      setCookies("lang").value must be_==("fr")
-      setCookies("logged").maxAge must beSome
-      setCookies("logged").maxAge.get must be_<=(-86000)
-      val playSession = Session.decodeFromCookie(setCookies.get(Session.COOKIE_NAME))
-      playSession.data.size must be_==(2)
-      playSession.data must havePair("user" -> "kiki")
-      playSession.data must havePair("langs" -> "fr:en:de")
-    }
-
-    "ignore session cookies that have been tampered with" in new WithApplication {
-      val data = Map("user" -> "alice")
-      val encodedSession = Session.encode(data)
-      // Change a value in the session
-      val maliciousSession = encodedSession.replaceFirst("user%3Aalice", "user%3Amallory")
-      val decodedSession = Session.decode(maliciousSession)
-      decodedSession must beEmpty
-    }
-
-    "support a custom application context" in {
-      "set session on right path" in new WithFooPath {
-        Cookies.decode(Ok.withSession("user" -> "alice").header.headers("Set-Cookie")).head.path must_== "/foo"
-      }
-
-      "discard session on right path" in new WithFooPath {
-        Cookies.decode(Ok.withNewSession.header.headers("Set-Cookie")).head.path must_== "/foo"
-      }
-
-      "set flash on right path" in new WithFooPath {
-        Cookies.decode(Ok.flashing("user" -> "alice").header.headers("Set-Cookie")).head.path must_== "/foo"
-      }
-
-      // flash cookie is discarded in PlayDefaultUpstreamHandler
-    }
-
-    "support a custom session domain" in {
-      "set session on right domain" in new WithFooDomain {
-        Cookies.decode(Ok.withSession("user" -> "alice").header.headers("Set-Cookie")).head.domain must beSome(".foo.com")
-      }
-
-      "discard session on right domain" in new WithFooDomain {
-        Cookies.decode(Ok.withNewSession.header.headers("Set-Cookie")).head.domain must beSome(".foo.com")
-      }
-    }
-
-    "support a secure session" in {
-      "set session as secure" in new WithSecureSession {
-        Cookies.decode(Ok.withSession("user" -> "alice").header.headers("Set-Cookie")).head.secure must_== true
-      }
-
-      "discard session as secure" in new WithSecureSession {
-        Cookies.decode(Ok.withNewSession.header.headers("Set-Cookie")).head.secure must_== true
-      }
-    }
-
-  }
-
-  abstract class WithFooPath extends WithApplication("application.context" -> "/foo")
-
-  abstract class WithFooDomain extends WithApplication("session.domain" -> ".foo.com")
-
-  abstract class WithSecureSession extends WithApplication("session.secure" -> true)
-
-  abstract class WithApplication(config: (String, Any)*) extends Around with Scope {
-    import play.api._
-
-    implicit lazy val app: Application = FakeApplication(Map("application.secret" -> "pass") ++ config.toMap)
-
-    override def around[T: AsResult](t: => T): SpecsResult = {
-      Play.start(app)
-      try {
-        AsResult(t)
-      } finally {
-        Play.stop()
-      }
     }
   }
